refactor: useBeforeUnload.tsを追加し、重複コードを整理
